Small Caching Fix, Icons, Debug Information and Methodology Fix
Changed how non-pastebin files are handled as to produce a .cctemp file
in the cache with a name that is the base 36 result of a crc32c
checksum. This is a much better solution then using "cc-auto-file".

Moved the setIcons method to Resources as a static method. Added better
debug information when loading and saving preferences. WebLink is a
Hyperlink wrapper that is functional on instantiation.

TypingMethodology had a BAD issue with in-line and multi-line comments.
This has been resolved through a new flag and logic fixes. Also added a
Pattern that removes unnecessary whitespace, condensing down to one
space.

GuiAccessor now has a openSite(String) method intended ot be used to
access a GUI Framework-dependent method of browsing to websites using
the Default Browser (i.e. HostServices or Desktop for JavaFX or Swing
respectively).

Other cleanup.