v3.0.0 Official Release -- Initial Minification and Polish
This update brings many bug fixes that will be described in the Change
Log. Many relate to GUI functionality and user-friendliness as well as
overall codebase maintainability. One of the most notable additions is
the "scripts" folder which contains any scripts used by the application
in operation, currently this is only 'luamin' which runs as a JavaScript
script under the Nashorn JavaScript engine provided in Java 8.

luamin allows for the minification of Lua source code. Notably it does
not work perfectly on all files.

Some tooltips have been added and some spelling mishaps have been fixed.
Additionally some of the text has been clarified.

The application will now also maintain a small cache of files in it's
working directory where any downloaded f iles are stored for ease of use
later in operation. We should probably add an uninstall script to the
application.

Due to the numerous bug fixes in this version, it was deemed necessary
to make a 3.0.0 major update release. It is that much better.