Post-final Fixes for Windows computers.
Mostly relating to slips of the mind when setting POSIX file permissions (i.e. all linux files are given the permissions "rw-rw-rw-" for ease of access by the program, most of these files are in the .cctyper-cache folder), there was no check on filesystem and so attempts to set POSIX permissions on a windows computer caused an exception. The check fixes this so the minifier experience on both windows and linux is identical.